+++ docs/CHANGELOG.md
@@ -57,6 +57,7 @@ Changelog
 - Chng: Update min HumHub version to v1.6.3
 - Chng: Major refactoring
 - Enh: Enable default permissions
+- Enh: Support RESTful API module
 
 1.0.15 - March 28, 2020
 -----------------------